+++ module/Shared/src/Application/Outbox/OutboxInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Shared\Application\Outbox;
+
+use ExtendsSoftware\ExaPHPExample\Shared\Domain\DomainEventInterface;
+
+interface OutboxInterface
+{
+    public function append(DomainEventInterface $event): void;
+}